Refactor: remove package name underscores
