+++ shared/src/commonMain/kotlin/data/credentials/EuPidCredentialAttributeCategorization.kt
@@ -1,7 +1,7 @@
 package data.credentials
 
 import at.asitplus.jsonpath.core.NormalizedJsonPath
-import at.asitplus.wallet.app.common.third_party.at.asitplus.jsonpath.core.plus
+import at.asitplus.wallet.app.common.thirdParty.at.asitplus.jsonpath.core.plus
 import at.asitplus.wallet.eupid.EuPidScheme
 import data.PersonalDataCategory
