+++ shared/src/commonMain/kotlin/data/credentials/HealthIdCredentialAttributeCategorization.kt
@@ -1,7 +1,7 @@
 package data.credentials
 
 import at.asitplus.jsonpath.core.NormalizedJsonPath
-import at.asitplus.wallet.app.common.third_party.at.asitplus.jsonpath.core.plus
+import at.asitplus.wallet.app.common.thirdParty.at.asitplus.jsonpath.core.plus
 import at.asitplus.wallet.healthid.HealthIdScheme
 import at.asitplus.wallet.healthid.HealthIdScheme.Attributes
 import data.PersonalDataCategory
