+++ shared/src/commonMain/kotlin/data/credentials/TaxIdCredentialAttributeCategorization.kt
@@ -2,7 +2,7 @@ package data.credentials
 
 import at.asitplus.jsonpath.core.NormalizedJsonPath
 import at.asitplus.jsonpath.core.NormalizedJsonPathSegment.NameSegment
-import at.asitplus.wallet.app.common.third_party.at.asitplus.jsonpath.core.plus
+import at.asitplus.wallet.app.common.thirdParty.at.asitplus.jsonpath.core.plus
 import at.asitplus.wallet.taxid.TaxIdScheme
 import data.PersonalDataCategory
