+++ shared/src/commonMain/kotlin/ui/composables/credentials/CredentialCardHeader.kt
@@ -3,9 +3,9 @@ package ui.composables.credentials
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import at.asitplus.wallet.app.common.third_party.at.asitplus.wallet.lib.agent.representation
-import at.asitplus.wallet.app.common.third_party.at.asitplus.wallet.lib.data.iconLabel
-import at.asitplus.wallet.app.common.third_party.at.asitplus.wallet.lib.data.uiLabel
+import at.asitplus.wallet.app.common.thirdParty.at.asitplus.wallet.lib.agent.representation
+import at.asitplus.wallet.app.common.thirdParty.at.asitplus.wallet.lib.data.iconLabel
+import at.asitplus.wallet.app.common.thirdParty.at.asitplus.wallet.lib.data.uiLabel
 import at.asitplus.wallet.lib.agent.SubjectCredentialStore
 import ui.composables.LabeledText
 import ui.composables.PersonAttributeDetailCardHeading
